requirements.txt
flask
flask_sqlalchemy
psycopg2-binary
python-dotenv
marshmallow
flask-marshmallow